Adds films endpoints, changes project structure
