+++ app/src/api/v1/films/router.py
@@ -1,84 +1,37 @@
+from typing import Annotated
 from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Query, status
 
-from films.services import FilmService, GenreService, PersonService
+from services import FilmService
+from .schemas import DetailedFilm, ShortenedFilm
 
-from .models import DetailedFilm, DetailedGenre, DetailedPerson, ShortenedFilm
+SORTING_PARAMS = ['imdb_rating:asc', 'imdb_rating:desc']
 
-router = APIRouter()
+router = APIRouter(prefix="/films", tags=['Films'])
 
 
-@router.get("/films/{film_id}")
+def sorting_params(sort: str = Query(None)):
+    if sort not in SORTING_PARAMS:
+        return None
+    return sort
+
+
+@router.get("/{film_id}")
 async def film_details(
     film_id: UUID = Path(...), film_service: FilmService = Depends(FilmService.get_instance)
 ) -> DetailedFilm:
     film = await film_service.get_by_id(film_id)
     if not film:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Film not found")
-
     return DetailedFilm.from_elastic_schema(film)
 
 
-@router.get("/films")
+@router.get("/")
 async def film_list(
-    search: str = Query(None),
-    sort: str = Query(None),
+    search: Annotated[str | None, Query(max_length=50)] = None,
+    sort: Annotated[str | None, Depends(sorting_params)] = None,
     film_service: FilmService = Depends(FilmService.get_instance),
 ) -> list[ShortenedFilm]:
     films = await film_service.get_many(sort_params=sort, search_params=search)
-    return [ShortenedFilm.from_elastic_schema(film) for film in films]
-
-
-@router.get("/genres")
-async def genre_list(genre_service: GenreService = Depends(GenreService.get_instance)) -> list[DetailedGenre]:
-    genres = await genre_service.get_many()
-    return [DetailedGenre.from_elastic_schema(genre) for genre in genres]
-
-
-@router.get("/genres/{genre_id}")
-async def genre_details(
-    genre_id: UUID = Path(...), genre_service: GenreService = Depends(GenreService.get_instance)
-) -> DetailedGenre:
-    genre = await genre_service.get_by_id(genre_id)
-    if not genre:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Genre not found")
-
-    return DetailedGenre.from_elastic_schema(genre)
-
-
-@router.get("/persons")
-async def person_list(
-    search: str = Query(None),
-    sort: str = Query(None),
-    person_service: PersonService = Depends(PersonService.get_instance),
-) -> list[DetailedPerson]:
-    persons = await person_service.get_many(sort_params=sort, search_params=search)
-    return [DetailedPerson.from_elastic_schema(person) for person in persons]
-
-
-@router.get("/persons/{person_id}")
-async def person_details(
-    person_id: UUID = Path(...), person_service: PersonService = Depends(PersonService.get_instance)
-) -> DetailedPerson:
-    person = await person_service.get_by_id(person_id)
-    if not person:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Person not found")
-
-    return DetailedPerson.from_elastic_schema(person)
-
-
-@router.get("/persons/{person_id}/films")
-async def person_films(
-    person_id: UUID = Path(...),
-    person_service: PersonService = Depends(PersonService.get_instance),
-    film_service: FilmService = Depends(FilmService.get_instance),
-) -> list[ShortenedFilm]:
-    person = await person_service.get_by_id(person_id)
-    if not person:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Person not found")
-
-    films = await film_service.get_many(
-        search_params=f"actors:{person_id} OR writers:{person_id} OR directors:{person_id}"
-    )
     return [ShortenedFilm.from_elastic_schema(film) for film in films]
