+++ app/src/main.py
@@ -5,6 +5,8 @@ import uvicorn
 from fastapi import FastAPI
 
 from api.v1.films.router import router as films_router
+from api.v1.genres.router import router as genres_router
+from api.v1.persons.router import router as persons_router
 from core.config import Settings
 from db import ElasticApp, RedisApp
 
@@ -15,7 +17,11 @@ app = FastAPI(
     openapi_url=settings.openapi_documentation_url,
     docs_url=settings.api_documentation_url,
 )
+
 app.include_router(films_router, prefix="/api/v1")
+app.include_router(genres_router, prefix="/api/v1")
+app.include_router(persons_router, prefix="/api/v1")
+
 redis = RedisApp(host=settings.redis_host, port=settings.redis_port)
 elastic = ElasticApp(host=settings.elastic_host, port=settings.elastic_port)
