+++ app/src/models/genre.py
@@ -1,0 +1,8 @@
+from uuid import UUID
+
+from core.mixins import CustomBaseModel
+
+
+class Genre(CustomBaseModel):
+    id: UUID
+    name: str
