+++ app/src/services/__init__.py
@@ -1,0 +1,3 @@
+from .film_service import FilmService
+from .genre_service import GenreService
+from .person_service import PersonService
