+++ app/src/services/film_service.py
@@ -1,11 +1,12 @@
 from typing import Optional
 from uuid import UUID
 
+import orjson
 from elasticsearch import NotFoundError
 
-from .constants import FILM_CACHE_EXPIRE, MOVIES_INDEX_NAME
+from core.constants import FILM_CACHE_EXPIRE, MOVIES_INDEX_NAME
+from models.film import Film
 from .mixins import ServiceMixin
-from .models import Film, Genre, Person
 
 
 class FilmService(ServiceMixin):
@@ -20,35 +21,24 @@ class FilmService(ServiceMixin):
 
     async def _get_film_from_elastic(self, film_id: UUID) -> Optional[Film]:
         try:
-            doc = await self.elastic.get(index=MOVIES_INDEX_NAME, uuid=film_id)
+            doc = await self.elastic.get(index=MOVIES_INDEX_NAME, id=film_id)
         except NotFoundError:
             return None
-        return Film(**doc["_source"])
+        return Film.load_model(doc["_source"])
 
     async def _film_from_cache(self, film_id: UUID) -> Optional[Film]:
         data = await self.redis.get(str(film_id))
         if not data:
             return None
-        return Film.model_validate(data)
+        return Film.model_validate(orjson.loads(data))
 
     async def _put_film_to_cache(self, film: Film):
-        await self.redis.set(str(film.uuid), film.model_dump_json(), FILM_CACHE_EXPIRE)
+        await self.redis.set(str(film.id), film.model_dump_json(), FILM_CACHE_EXPIRE)
 
     async def get_many(self, sort_params=None, search_params=None) -> list[Film]:
-        pass
-
-
-class GenreService(ServiceMixin):
-    async def get_by_id(self, genre_id: UUID) -> Optional[Genre]:
-        pass
-
-    async def get_many(self, **kwargs) -> list[Genre]:
-        pass
-
-
-class PersonService(ServiceMixin):
-    async def get_by_id(self, person_id: UUID) -> Optional[Person]:
-        pass
-
-    async def get_many(self, **kwargs) -> list[Person]:
-        pass
+        docs = await self.elastic.search(
+            index=MOVIES_INDEX_NAME,
+            query={"match": {"title": {"query": search_params}}},
+            sort=sort_params
+        )
+        return [Film.load_model(doc["_source"]) for doc in docs["hits"]["hits"]]
