+++ app/src/services/mixins.py
@@ -1,21 +1,13 @@
 from typing import Self
 
-from elasticsearch import AsyncElasticsearch
-from fastapi import Depends
-from redis.asyncio import Redis
-
 from core.mixins import Singleton
 from db import ElasticApp, RedisApp
 
 
 class ServiceMixin(Singleton):
-    def __init__(
-        self,
-        redis: Redis = Depends(RedisApp.get_instance),
-        elastic: AsyncElasticsearch = Depends(ElasticApp.get_instance),
-    ) -> None:
-        self.redis = redis
-        self.elastic = elastic
+    def __init__(self):
+        self.redis = RedisApp.redis_client()
+        self.elastic = ElasticApp.elastic_client()
 
     @classmethod
     def get_instance(cls) -> Self:
